Extracted action handler registry from diagram server
